+++ back-end/.gitignore
@@ -2,6 +2,11 @@
 /venv
 
 .idea/
-API_KEYS.py
+../API_KEYS.py
 .ebignore
 .elasticbeanstalk/
+
+# Elastic Beanstalk Files
+.elasticbeanstalk/*
+!.elasticbeanstalk/*.cfg.yml
+!.elasticbeanstalk/*.global.yml
