+++ back-end/app/models.py
@@ -1,14 +1,4 @@
-
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_cors import CORS
-from API_KEYS import *
-
-
-app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = db_url
-CORS(app)
-db = SQLAlchemy(app)
+from app import db
 
 class User(db.Model):
     __tablename__ = "User"
