+++ back-end/requirements.txt
@@ -19,7 +19,7 @@ Jinja2==3.0.2
 MarkupSafe==2.0.1
 pafy==0.5.5
 protobuf==3.19.1
-psycopg2==2.9.2
+psycopg2-binary==2.9.2
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
 PyJWT==2.3.0
